Moved Header and Footer to their own components
Moved Header and Footer to their own components to prevent weird loading issues with Gridsome. Also updated Gridsome to the latest version
