+++ web/src/pages/Projects.vue
@@ -50,18 +50,6 @@
           }
           caption
           alt
-          hotspot {
-            x
-            y
-            height
-            width
-          }
-          crop {
-            top
-            bottom
-            left
-            right
-          }
         }
       }
     }
